Made necessary changes, put drawing code back into main
