+++ 3D Graph/3D Graph.vcxproj.filters
@@ -39,7 +39,7 @@
     <ClInclude Include="Vector2D.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="MainLoop.h">
+    <ClInclude Include="ControlLoop.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
@@ -62,7 +62,7 @@
     <ClCompile Include="Vector2D.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="MainLoop.cpp">
+    <ClCompile Include="ControlLoop.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
